FEATURE - Add entities mappers
Se agregaron los mappers para las entities de Conversion, CurrentExchange y ConversionsHistory en sus respectivos módulos de history y calculator
