+++ app/src/main/java/com/varqulabs/dolarblue/calculator/data/local/database/entities/CurrentExchangeRateEntity.kt
@@ -3,7 +3,6 @@ package com.varqulabs.dolarblue.calculator.data.local.database.entities
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.varqulabs.dolarblue.history.domain.model.CurrentExchangeRate
 import kotlinx.datetime.LocalDateTime
 
 @Entity(tableName = "current_exchange_rate_table")
@@ -13,11 +12,4 @@ data class CurrentExchangeRateEntity(
     @ColumnInfo(name = "pesosBob") val pesosBob: String,
     @ColumnInfo(name = "pesosArg") val pesosArg: String,
     @ColumnInfo(name = "date") val date: LocalDateTime,
-)
-
-fun CurrentExchangeRateEntity.mapToModel() = CurrentExchangeRate(
-    id = id,
-    pesosBob = pesosBob,
-    pesosArg = pesosArg,
-    date = date
 )
