+++ app/src/main/java/com/varqulabs/dolarblue/history/data/local/database/entities/relations/ConversionsHistoryRelation.kt
@@ -4,8 +4,6 @@ import androidx.room.Embedded
 import androidx.room.Relation
 import com.varqulabs.dolarblue.calculator.data.local.database.entities.ConversionEntity
 import com.varqulabs.dolarblue.calculator.data.local.database.entities.CurrentExchangeRateEntity
-import com.varqulabs.dolarblue.calculator.data.local.database.entities.mapToModel
-import com.varqulabs.dolarblue.history.domain.model.ConversionsHistory
 
 data class ConversionsHistoryRelation(
     @Embedded val currentExchangeRate: CurrentExchangeRateEntity,
@@ -16,7 +14,3 @@ data class ConversionsHistoryRelation(
     val conversions: List<ConversionEntity>
 )
 
-fun ConversionsHistoryRelation.mapToModel() = ConversionsHistory(
-    currentExchangeRate = currentExchangeRate.mapToModel(),
-    conversions = conversions.map { it.mapToModel() }
-)
